pyproject.toml
[project]
name = "deepsearchjobs"
version = "0.1.0"
description = "DeepSearchJobs is a module of Play2Path dedicated to fetching unique job opportunities across the web, helping you discover and apply to positions that best match your profile, increasing your chances of success."
authors = [
    {name = "abdelwakil.benabdi@play2path.com"}
]
license = {text = "CC BY-NC 4.0"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "psycopg (>=3.3.2,<4.0.0)",
    "beautifulsoup4==4.13.3",
    "openai (>=2.15.0,<3.0.0)",
    "urllib3==2.3.0",
    "python-dotenv==1.0.1",
    "lxml==5.3.1",
    "lxml-stubs==0.5.1",
    "python-Levenshtein==0.27.1",
    "sentence-transformers==4.0.2",
    "hf_xet",
    "playwright (>=1.57.0,<2.0.0)",
    "playwright-stealth (>=2.0.1,<3.0.0)",
    "requests==2.32.5",
    "mypy==1.18.2",
    "torch (==2.9.0)",
    "redis (==7.0.1)",
    "types-pika (>=1.2.0b1,<2.0.0)",
    "types-requests (>=2.32.4.20250913,<3.0.0.0)",
    "aio-pika (>=9.5.8,<10.0.0)",
    "aioboto3 (>=15.5.0,<16.0.0)",
    "psycopg-pool (>=3.3.0,<4.0.0)",
    "types-aioboto3 (>=15.5.0,<16.0.0)",
    "markdownify (>=1.2.2,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "python-pptx (>=1.0.2,<2.0.0)",
    "pandas-stubs (>=3.0.0.260204,<4.0.0.0)",
    "pymupdf (>=1.27.1,<2.0.0)",
    "pymupdf4llm (>=0.3.4,<0.4.0)",
    "types-redis (>=4.6.0.20241004,<5.0.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"
[tool.poetry.dependencies]
torch = {source = "pytorch"}

[tool.poetry]
package-mode = false